Player Settings sepparation
